+++ openapi/openapi.cabal
@@ -1,0 +1,157 @@
+name:
+  openapi
+version:
+  0.0.0.1
+synopsis:
+  OpenAPI schemas
+description:
+  Implements the OpenAPI 3.0.3 spec
+bug-reports:
+  https://github.com/felixmulder/servant-openapi
+cabal-version:
+  1.24
+license:
+  BSD3
+license-file:
+  LICENSE
+author:
+  Felix Mulder
+maintainer:
+  felix.mulder@gmail.com
+copyright:
+  Felix Mulder
+category:
+  Web, OpenAPI
+build-type:
+  Simple
+extra-source-files:
+    README.md
+  , CHANGELOG.md
+
+source-repository head
+  type: git
+  location: git://github.com/felixmulder/servant-openapi.git
+
+library
+  default-extensions:
+    BlockArguments
+    DataKinds
+    DeriveAnyClass
+    DeriveGeneric
+    DerivingStrategies
+    DerivingVia
+    DuplicateRecordFields
+    EmptyDataDeriving
+    FlexibleContexts
+    FlexibleInstances
+    GeneralizedNewtypeDeriving
+    KindSignatures
+    LambdaCase
+    MultiParamTypeClasses
+    NamedFieldPuns
+    OverloadedStrings
+    PolyKinds
+    RankNTypes
+    ScopedTypeVariables
+    StandaloneDeriving
+    TupleSections
+    TypeApplications
+    TypeApplications
+    TypeApplications
+    TypeOperators
+    ViewPatterns
+
+  ghc-options:
+    -Wall
+    -Wredundant-constraints
+    -Wmissing-deriving-strategies
+    -fhide-source-paths
+
+  exposed-modules:
+    OpenAPI.Class
+    OpenAPI.Internal.Class
+    OpenAPI.Internal.Types
+
+  build-depends:
+      aeson                     >= 1.4.0    && < 1.5
+    , aeson-deriving
+    , base                      >= 4.9      && < 5
+    , containers                >= 0.6      && < 0.7
+    , unordered-containers
+    , generic-lens              >= 1.2      && < 1.3
+    , lens
+    , http-media                >= 0.8      && < 0.9
+    , lens                      >= 4.18     && < 4.19
+    , servant                   >= 0.16     && < 0.18
+    , text                      >= 1.2      && < 1.3
+    , yaml
+    , extra
+
+  hs-source-dirs:
+    src
+
+  default-language:
+    Haskell2010
+
+test-suite openapi-tests
+  type:
+    exitcode-stdio-1.0
+
+  default-language:
+    Haskell2010
+
+  main-is:
+    Main.hs
+
+  ghc-options:
+    -Wall -Wredundant-constraints -fhide-source-paths -threaded
+
+  hs-source-dirs:
+    test
+
+  other-modules:
+
+  default-extensions:
+    BlockArguments
+    DataKinds
+    DeriveAnyClass
+    DeriveGeneric
+    DerivingStrategies
+    DerivingVia
+    DuplicateRecordFields
+    EmptyDataDeriving
+    FlexibleContexts
+    FlexibleInstances
+    GeneralizedNewtypeDeriving
+    KindSignatures
+    LambdaCase
+    MultiParamTypeClasses
+    NamedFieldPuns
+    OverloadedStrings
+    PolyKinds
+    RankNTypes
+    ScopedTypeVariables
+    StandaloneDeriving
+    TupleSections
+    TypeApplications
+    TypeApplications
+    TypeApplications
+    TypeOperators
+    ViewPatterns
+
+  build-depends:
+      servant-openapi
+    , openapi
+
+    , aeson                     >= 1.4      && < 1.5
+    , aeson-deriving
+    , base                      >= 4.9      && < 5
+    , bytestring                >= 0.10     && < 0.11
+    , case-insensitive          >= 1.2      && < 1.3
+    , hedgehog                  >= 0.6      && < 1.1
+    , http-client               >= 0.6      && < 0.7
+    , http-media                >= 0.8      && < 0.9
+    , http-types                >= 0.12     && < 0.13
+    , string-conversions        >= 0.4      && < 0.5
+    , text                      >= 1.2      && < 1.3
+    , yaml                      >= 0.11.2.0 && < 0.12
